Fix: wand of opening zapped at statue brought up tipping menu
This bug may date back to 2020. I added the ability for statues to spill
their contents when zapped with opening in 88a00ef (spring 2020),
calling tipcontainer() to spill the contents, but then didn't change
that code when I refactored out dump_container() in 12b2dcc (fall 2020).
As a result, zapping statues would prompt the player with a menu for
where they wanted to tip the items, as if they were carrying the statue,
even if there were no items to spill out.

The bug might be more recent, with an upstream change that affected
tipcontainer(). Anyway, I fixed it by changing the statue-spill calls to
dump_container, which doesn't have any prompts.

A somewhat unrelated addition is that if you see items spill from the
statue, the wand gets identified, since only opening will do that.