Fix: statue traps didn't interact with wand of opening
This was weird given that the statue spilling its inventory is
presumably an interaction that affects it enough to trigger a statue
trap, so this makes it so that the opening beam does trigger the trap.

I decided that it would be an additional nice benefit if the items
spilled *before* the statue animated, so instead of going into the
monster's inventory the items end up on the floor. Granted, this is not
likely to matter much, because statue-trap statues don't contain the
monster's normal starting inventory.