Fix: Mini Maze themed room not generating properly
Noticed when testing it out that it was generating the maze walls and
connecting the room to corridors weirdly. I ended up deciding that using
des.mazewalk() here is bending it too far beyond its intended use. So I
replaced the call to that with a simple Lua mazewalk algorithm. This
gets rid of the weird double wall and the weird x and y placement math
that relied on ROWNO and COLNO. (Thus also allowing me to remove the
FIXME comment about that.)

Without the double wall, the room was 2 spaces shorter in both x and y.
I reclaimed that space in the x-dimension by extending the maze, but
didn't do the same in the y-dimension because tall rooms more often lead
to uninteresting levels.