Fix: monster hiding under something but on a trap
I observed this as a sanity check failure when testing a themed room.
The themed room specified a random monster, and also allowed regular
room filling with random contents. That generated a random trap, which
happened to be on the same square as the monster. The monster was a cave
spider, so it had generated with an object to hide under and had hidden
under it. The trap's later generation didn't do anything about that.
This wouldn't happen in regular room filling, since traps generate
before monsters, but is possible in themed rooms.

Fix it by unhiding any monster on a trap's space immediately after the
trap is placed.