Fix: don't show "Normally made of" in obj lookup for spellbooks
It was possible to discern some information about the true identities of
spellbooks: by querying object lookup for each spellbook, the player
could figure out which two spells would be "parchment" and "vellum"
since object lookup would only say "Normally made of leather." for those
spellbooks.

There were already some classes defined for specifically not showing
this because it leaks data; this just adds SPBOOK_CLASS to that
condition. Because it uses the same code, the "Normally made of" will be
shown for spellbook types whose appearance is already identified.