Give iron doors proper defsyms and tiles
This is a long-deferred feature from vanilla merge 83e664f, where in the
face of so much effort spent welding xNetHack changes with upstream
display code changes, no more was left for iron doors. They remained a
carve-out where the color of the glyph was computed on the fly.

This came to my attention when the fuzzer crashed on the "no visible
door" impossible (happens when detecting a secret door and the glyph for
a door flashes briefly on its space before it's converted to a normal
door). Getting rid of the on-the-fly color change code by adding iron
door defsyms was the most time-consuming way to fix that, but it's the
right thing to do.

The big beneficiary of this change is tileset players, who previously
had no way other than farlooking to differentiate between wooden doors
and "iron" doors that used the same wooden-looking tile. In creating the
iron door tiles, I mostly just palette swapped the wooden colors for the
basic colors in the iron bars tile, but there are a few less accent
pixels.

As of this commit, other colored walls for special rooms and Valley of
the Dead monochroming are still on-the-fly color changes in
map_glyphinfo. Other walls might be added at some later point (depends
on where exactly we end up going with special rooms vs themed rooms, and
if the willpower is there to make tile variants for all of them) but the
monochroming is unlikely to ever get a tile-friendly implementation.