Fix: mselftouch didn't check for dead monster
This could result in an indefinite recursion of extract_from_minvent ->
mselftouch -> minstapetrify -> xkilled -> monstone ->
extract_from_minvent -> ...

With a dead monster check, mselftouch won't be called recursively
because the first xkilled or monstone call will set mhp to 0.

While here, also fix a separate bug introduced in the same change where
the player was always credited for the kill. The fuzzer crash that found
this involved a monster-vs-monster case.