Fix: mdig_tunnel wasn't returning TRUE for digging monster's death
Apparently in the change which excised mb_trapped for doors, so too was
any code path which returned true from this function.

While I was figuring out where exactly to return true from it, I also
decided to restructure this bit of code. Instead of one large if
statement for the monster not being dead, it merely returns true early
if the monster is dead. It also checks if interacting with a door trap
made the door no longer closed, in which case the digging attempt is
aborted.