Fix: storm giant lightning spell had no handling for melee version
Reported by elunna. The effect of this was that you would just get "The
storm giant casts a spell!" with no further message and take the full
damage even if shock resistant, because AD_ELEC monster spells weren't
handled when I first introduced them

Fix it based on the fix from NerfHack: not too different from the other
spells, but also including a ugolemeffects() call. I'm not sure why that
isn't called for the AD_FIRE spell - perhaps because that is unused.