Add parameter to lifesaved_monster() that suppresses messages
An incoming bug fix requires calling lifesaved_monster, but printing any
messages from it wouldn't make sense because the life saving is not
happening on the current level. Add this offlevel option to retain all
the mechanical stuff while avoiding showing messages.

Note that wary_dog can still print messages, but it's enough of a corner
case not to worry about right now.