Fix: monster falling into air might lifesave and stay hovering
Noticed when fixing another bug that if the hero causes the death of a
monster in open air, which then gets lifesaved, the lifesaving would
happen in xkilled() and the monster would survive on the current
level... in midair with nothing holding it up, after the "plummet"
message.

Fix this by attempting to life-save the monster, using up any lifesaving
source it might have, before calling either mongone or xkilled.