Fix: "Thin long horizontal room" could generate as invalid shop
Shops require a line of empty spaces adjacent and perpendicular to the
door, and place items on the remaining spaces. In this particular themed
room, the room could generate as only 1 space high, and if it became a
shop then that 1 row would be the empty space and there would be nowhere
to put the items.

Fix this by making the room type "themed" when it is only 1 space high,
blocking any special rooms from generating. When it is 2 spaces high it
can still turn into a shop or other special room.