Fix: archfiend bribe() could get called when hero had no gold
Fuzzer identified an edge case that fell through the cracks and hit the
impossible in money2mon: if an archfiend demanded at least one item from
you and you turned it over, then had no cash, it would attempt to get a
bribe and transfer 0 zorkmids to the monster.

It doesn't appear that bribe is intended to be called with 0 gold since
vanilla checks for this everywhere it's called. So I didn't add any
guard case into bribe() itself; the fix is just to insert an additional
pline noting you have no money instead of calling bribe().