Fix: cold trap reduced mon max HP and not current after life-save
An assumption stated in the comment that updating mhp would not be
necessary turned out to be false: when the monster was killed and
life-saved, it returns false from thitm().

Since cold traps parallel fire traps, I used the same method of simply
checking and updating mhp. This is a bit strange, because it means a
monster that got killed by the trap and life-saved will still lose
maximum HP, but it's at least consistent now.