Fix: hellfill prefabs could still overlap
Rarer than I think this used to be, but the fuzzer still caught it - the
"spiral" prefab overwrote the edge of the "Kraken" prefab and turned
water into floor, resulting in an eel hiding out of water.

Fix this by enforcing that there can never be a prefab in the center and
another one in the half-left/right position at the same time. A
consequence of this is no level will have 3 prefabs anymore. There can
still be 2, but only about 30% of the time.

This will cause the same problems if prefabs get sufficiently wide to
start overlapping again from half-left and half-right positions, but
there are no plans at present to add any more super-wide ones.