Fix: trap ammo didn't have any weight set
The weight was uninitialized (or simply set to its default by mksobj()
of the first one) and meant that disarming the trap would give you the
correct quantity of ammo, but the weight of which could be incorrect.

Fix this by explicitly setting the weight right after setting the
quantity.