Remove lingering traces of hilite_hidden_stairs
This was removed as an option some time ago, but there were still a
couple bits of code sticking around involving that system.

As part of this, MG_PEACEFUL is shifted back one bit (tested it,
peacefuls are still highlighted appropriately, no changes).

Also, I noticed a comment in map_glyphinfo was somewhat out of date, so
I updated it to reflect what it currently does and doesn't cover.