Adjust all role-specific legacy strings to be plural
I noticed this for Healer and thought "the School of [deity]" kind of
implies they are not that impressive of a god if they have only one
school. Ditto for the other two roles that had a singular Thing of
Deity. Make them all plural.

I don't know Japanese, but after checking it appears that "Emakimono"
is correct as both a singular and plural noun, and "Emakimonos" would be
wrong. So no change there.