Fix: probing wand didn't give message or auto-identify
This got accidentally reverted probably in the most recent vanilla
merge, when vanilla's variable changes to engraving with wands collided
with xNetHack auto-identification of wands. Most appear to have been
merged properly, but the wand of probing case which was moved elsewhere
appears to have simply been reverted. That wasn't intended, so this
restores the behavior.