Consolidate "does object display a +/-" in a macro
There were 3 places existing in the code that checked for "is weapon or
armor or weptool or charged ring", and I have plans to add a 4th. So I
moved these cases into a new macro, spe_means_plus(). Didn't touch any
cases where this might technically have been ok to use but would amount
to "checking for charged ring where we've already established the object
is a ring". oc_charged is still fine in such cases.