Make restore ability cure several status ailments
This is aimed at further raising restore ability into the class of
potions that are useful to keep around, while justifying it as
"restoring your ability to see/walk/etc".

The potion was already distinct from unicorn horns after unicorn horns
were nerfed to not heal stats, but they now effectively give a
guaranteed cure of confusion, blindness, hallucination, and maybe
illness where a unicorn horn may randomly fail, and they cure wounded
legs which unicorn horns won't fix.

The spell of restore ability, since it uses the same code, also got
buffed by this change, but it can practically be used to cure only
wounded legs and hallucination. If you are confused or stunned, you
couldn't have cast it, and illness won't be cured by it (since that
would make it a better form of cure sickness).

Also in this commit is adjusting the restore ability effect when thrown
at monsters: previously it inexplicably healed them to maximum HP
despite not being a healing potion, now it cures their blindness,
confusion, and stunning.