Fix: double Bell of Opening in Val and Arc quests
This is the result of the fix in 2d4f989 that makes a nemesis who's
about to receive lua-specified inventory drop any special items they may
have already received. In the commit note there I even noted two Bells
could generate if a level developer specified the nemesis gets the Bell
and fails to specify keep_default_invent = true, but also noted that
"two Bells is better than none", which it is - the player who reported
this wasn't facing an unwinnable situation.

Specifying keep_default_invent = true as I did here was not necessary to
fix the two-Bells issue, but does address another problem: that the
default behavior when keep_default_invent is false is to *drop* special
items, meaning the nemesis would either abandon the Bell on the floor or
waste turns warping back to pick it up once they moved off. I considered
doing a more comprehensive fix that would keep the special items in the
nemesis's inventory, but decided not to since it would duplicate code
from mdrop_special_objs.