Fix: blessed gold detection could use up the scroll twice
This is an oversight introduced in commit 9cd266ae, which extended the
blessed scroll of gold detection to also find gems after it finds gold.
The issue is that both gold_detect and object_detect, if passed an
object pointer, might unconditionally delete the object (another
instance of unintuitive behavior where the deletion is hidden way down
in strange_feeling). So calling them twice on the same object pointer
is incorrect, but also avoiding gem detection if gold detection fails to
find anything is incorrect.

I resolved this by calling object_detect with a null pointer since it
didn't do anything special anyway when passed a blessed scroll of gold
detection. Also added another strange_feeling call but also with a null
pointer so it will simply deliver its message and not delete anything.

For posterity's sake this is the first crash detected in Junethack 2025.