Designate the Wizard's Tower branch as hellish
Someone reported a while back that one of the possible ways to get past
the Wizard's Tower puzzle was simply levelporting up to the top level.
This was an oversight during the Gehennom redesign; it also meant that
various other effects like the odds of demon gating, foocubi being
more powerful, and the hero being able to pray acted as if they weren't
in Gehennom.

Prior to the tower being pulled out into its own branch, it was treated
as part of Gehennom, so this reestablishes that, putting in all the
behaviors associated with Gehennom levels.

Note that Vlad's Tower remains non-hellish, and the non-Gehennom
behaviors remain in effect there.