Fix: charm monster wasn't being handled in spelleffects()
Last December PatR moved a case statement for the spell around to use
the blessed taming scroll effect when charm monster cast at Skilled or
Expert, but somehow the xNetHack copy of the code deleted the case
entirely. The large vanilla merge in March seems like the place this
would have happened but I can't find it in the commit diff, nor can I
find any notes indicating I intended to change the behavior, so I reset
it to vanilla's version of the code.