Fix: shopkeepers bought unidentified gems for lots of money
As discovered by KarlMHanson on Github, this is not the gem being sold
for its natural price - it's the gem being sold with the GEMSTONE
material price multiplier added. The special case code for unidentified
gems is intended to reset the multiplier and divisor to wipe out things
that normally influence the sell cost and make the gem sell for only a
few gold pieces - but it only actually set divisor to 1 because that's
the only variable that could change prior to the gem special case in
vanilla.

Rather than the suggested solution of wrapping the object material
adjustment in a (oclass != GEM_CLASS) block, I found it simpler to reset
the multiplier to 1 in the gem special case.

Fixes #224