Fix: thiefstone couldn't be re-keyed
When I added extra uses for dipping objects in potions of restore
ability, I included an early return for that block. This meant the
thiefstone re-keying code which was left below that could never be
reached with a non-cursed potion of restore ability, meaning effectively
that thiefstones could no longer be re-keyed.

Fix is to move the re-keying code into the general case for restore
ability. While here, I slightly adjusted the undo-negative-enchantment
restore ability behavior to hit poof() -> trycall() rather than directly
calling trycall() itself (fixing a minor bug; the player would have been
prompted to call the potion even if its dknown wasn't set).

I decided to leave the thiefstone-rekeying case as something that will
prompt a trycall rather than automatically identifying the potion, since
it might not be obvious that it was restore ability even if the player
has identified the thiefstone and can see it went from inactive to
keyed.

There is no longer a case that handles re-keying a thiefstone with a
cursed potion of restore ability. Previously it printed a unique message
and used up the potion. Now it falls into the general dipping case,
meaning it results in "Interesting..." and doesn't use up the potion.

Fixes #226