Make Arc quest guardians ask for potions of object detection
There was an existing line asking the player if they had any potions of
hallucination for sale; that line isn't changed, but is now also
duplicated for potions of object detection, since that potion would of
course be hugely useful to archeologists.

This was a stray minor feature from the Arc quest redesign that didn't
make it in.