Fix: Abyss levels contained no empty air
This was due to using a negative y-coordinate to anchor the ellipse,
which used to work but there was some change introduced in the last
vanilla merge that means they are translated into an unsigned char at
some point and resulted in the y-coordinate being around 250. This meant
the ellipse selection was off the map and thus no terrain got replaced
by air.