Remove bonus Luck point for throwing named valuable gems at unicorns
This is one of NetHack's features that relies on most players being
unspoiled, and it doesn't work out as well in the age where being
spoiled is the most common way to play. It was optimal and tedious to
type-name all gems before throwing them at a coaligned unicorn, knowing
that by the rules you would get an extra point of Luck out of it for
"making guesses" if it turned out to be valuable. This was regardless of
what the name actually was; even if you were just naming gems "red gem
1", "red gem 2" etc for bookkeeping and were not guessing at all, the
game would still award the bonus.

The main debate I had here was whether, instead of removing the luck
bonus, I should narrow it to checking specifically for a correct guess.
I.e. do a start-of-string match on the player-named string and the gem's
correct name (start of string so that if the player names a gem e.g.
"jet stone" when the defined name of the gem in objects.h is just "jet"
it still matches). The reason I decided against this is because it would
just shift the tedious optimal behavior to a more annoying one: now the
optimal choice would be to type-name any unknown gem with a real gem
name before throwing it (temporarily erasing any actual deductive notes
you were storing in the discoveries list), and hope you're correct.