Reduce chances of silver and mithril on iron-base items
This is based on a couple reports and confirmation from playtesting -
silver and mithril items were a lot more common than one would expect;
since so many items are iron, 1 in 20 of them being silver meant you saw
quite a bit of silver (and to a lesser extent, the same for mithril).

While I do intend to have silver more common in xNetHack than it is in
vanilla (where your first glimpse of it is likely to be a watch
captain's saber in Minetown), I also intend it and mithril to be rare
enough so that a character won't have necessarily have a silver weapon
and a mithril armor piece by the end of the Mines.