Remove getmap shim from Arc-filb
This shim was added in 2021 because at the time, nh.getmap() always
returned absolute coordinates, resulting in an off-by-one error when
this level script used it because typically, the "actual" map starts at
x=1 unless defined otherwise.

That was then fixed in a 2022 patch I submitted upstream, which has
subsequently been merged back to xNetHack.

This commit removes the now-unneeded shim as well as a commented out
testing line. It does NOT fix the borked Arc-filb generation - the cause
of that is an upstream selection.match bug, which I am not sure of the
proper fix (issue #1467 in the upstream nethack repo).