Fix: thiefstone put items into a bag of tricks
Bags of tricks for some reason return true from Is_container despite not
actually being usable as one. This meant that thiefstones could teleport
items into a bag of tricks, which could then not easily be gotten out
(it was possible, via wand of opening or polymorphing the bag, but this
was certainly not intended in the first place).

Adding an exception so bags of tricks don't count as containers breaks
them in several ways, however, so instead I changed thiefstone behavior
such that they now ignore bags completely. It's cleaner this way (rather
than special-casing an exception only for bags of tricks) and slightly
more thematic; sending items to a treasure chest makes more sense than a
"treasure bag".