Fix: archfiends would cast entomb and dark speech while peaceful
This was reported as a bug a while ago, and I recently confirmed it -
the archfiend needed to be "activated" and on the map to do it, but once
they were they would start hitting you with spells designed to be
offensive. The special spell selection code I added just chose a spell
without regard for them being peaceful, so this fixes it by treating all
their offensive spell selections as "useless" if they are peaceful.